LoginDao fix

+++ src/pl/workreporter/web/beans/security/login/LoginDao.java
@@ -9,10 +9,10 @@ import java.util.List;
  */
 public interface LoginDao {
     void createAccount(String login, String password, String tempPassword, String email, int status);
-    String getEmail(int id);
+    String getEmail(long id);
     String getEmail(String login);
-    String getLogin(int id);
-    String getPasswordHash(int id);
+    String getLogin(long id);
+    String getPasswordHash(long id);
     String getPasswordHash(String login);
     String getPasswordHashByEmail(String email);
     CompleteUserDetails loadUserDetails(String login);
